+++ RigidBody/game.ts
@@ -10,7 +10,7 @@ import {sys_framerate} from "./systems/sys_framerate.js";
 import {sys_lifespan} from "./systems/sys_lifespan.js";
 import {sys_light} from "./systems/sys_light.js";
 import {sys_physics} from "./systems/sys_physics.js";
-import {sys_render} from "./systems/sys_render.js";
+import {sys_render} from "./systems/sys_render1.js";
 import {sys_transform} from "./systems/sys_transform.js";
 import {World} from "./world.js";
