+++ RigidBody/scenes/sce_stage.ts
@@ -2,7 +2,7 @@ import {blueprint_camera} from "../blueprints/blu_camera.js";
 import {collide} from "../components/com_collide.js";
 import {control_spawner} from "../components/com_control_spawner.js";
 import {light_directional} from "../components/com_light.js";
-import {render_diffuse} from "../components/com_render_diffuse.js";
+import {render_diffuse} from "../components/com_render1.js";
 import {rigid_body} from "../components/com_rigid_body.js";
 import {instantiate} from "../core.js";
 import {Game, Layer} from "../game.js";
