+++ RigidBody/systems/sys_render1.ts
@@ -1,0 +1,1 @@
+../../core/systems/sys_render1.ts
