+++ RigidBody/world.ts
@@ -3,7 +3,7 @@ import {Collide} from "./components/com_collide.js";
 import {ControlSpawner} from "./components/com_control_spawner.js";
 import {Lifespan} from "./components/com_lifespan.js";
 import {Light} from "./components/com_light.js";
-import {Render} from "./components/com_render.js";
+import {Render} from "./components/com_render1.js";
 import {RigidBody} from "./components/com_rigid_body.js";
 import {Transform} from "./components/com_transform.js";
